refactor: refactor typography components and styles for improved structure and consistency
