+++ src/atomic/atm.typography/index.ts
@@ -1,3 +1,3 @@
-import { Text, InputCaption, InputLabel, InputValue, Link } from './typegraphy.component';
+import { Text, InputCaption, InputLabel, InputValue, Link } from './typography.component';
 
 export { InputCaption, InputLabel, InputValue, Link, Text };
